This code is repetitive (i spelt that wrong didnt i)
I've gotten down the basics of what I want the game to do. Like the scores, the checks, boxes generating random numbers, all of that. But it doesn't work the way it's supposed to. For example the code repeats itself more than it has to, the scores don't add, and there's no limit to how many times you can input one number. So there is a lot to fix